Remove (now) unneeded compaction mutator counter for completion
We use a counter, which is incremented before a mutator start compacting
a moving-space page, and decrements once done. On the other side,
GC-thread, after finishing sequentially compacting all the moving-space
pages, waits for the counter to reach 0 before uffd-unregistering the
moving-space and madvising the from-space.

This is no longer required as now we unregister moving-space after
finishing linear-alloc space, which also uses a similar counter,
which is sufficient for both now.

Bug: 320478828
Test: art/test/testrunner/testrunner.py --host
Change-Id: Ibd7f4306b007e1965567ae667288e60852f6da53