Allow RunCheckpoint to lock mutator repeatedly
Add an option to RunCheckpoint to acquire the mutator lock separately
for each suspended-thread checkpoint function we run. This avoids
keeping the initiating thread runnable, and unable to respond to suspend
requests. for the entire duration of dump generation. Otherwise that can
generate substantial jank, and suspend timeouts.

Modify ThreadList::Dump to use this version.

Replace the ScopedObjectAccess in DumpCheckpoint::Run() with an
assertion that we already hold the mutator lock.

This largely reverts aosp/238341. We believe that the deadlocks
motivating that were already made impossible by the more recent thread
suspension rewrite.

This weakens undocumented semantics of RunCheckpoints, and instead
documents the properties we actually do guarantee. Previously we
guaranteed that every thread in existence the the RunCheckpoint call
either ran the checkpoint, or had an ancesstor that did. We do not
believe this was ever relied upon.

Bug: 333708482
Test: Treehugger
Change-Id: Ief098b89e96bb647914bca4e467bca15971553cd