Revert "Clean up descriptor and member name comparisons."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c8f98d64-2a52-4d84-ae27-390f85e01474). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I247c47372b2740547d363237eae86f61dee0211b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com