Revert "Clean up descriptor and member name comparisons."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/44930bb0-5b9f-4791-9672-75f6c650cc57). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6edf8671fdb39b273c6782d3249d0be84b8ac87e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com