Serialize mutations to the global state of Pre-reboot Dexopt.
There is a global state that Pre-reboot Dexopt mutates, including
mounts and staged files. Before this change, it was up to the caller of
Pre-reboot Dexopt to make sure that different runs don't step on each
other's toes, and there was a shell command that can potentially mess up
with job runs initiated by the job scheduler. This was too fragile.

After this change, Pre-reboot Dexopt only exposes 3 APIs:
PreRebootDexoptJob.onStartJob, PreRebootDexoptJob.onStopJob,
PreRebootDexoptJob.onUpdateReady. The implementation of those APIs makes
sure that mutations to the global state are serialized.

This change paves the way for implementing telemetry, where the stats
becomes a part of the global state and mutations to it need to be
serialized.

Bug: 311377497
Test: atest ArtServiceTests
Change-Id: I5bdef52e15dd465edb3eb63b4b399bdc4860a8d4