Fixed missing end time on previous runs.
Before this change, there was a bug with metrics reporting. When a job
is rerun, the end time of previous runs are unexpectedly cleared.

When a job is cancelled and rerun, `recordJobEnded` is called by a
reporter inside chroot, and then `recordJobStarted` is called by a
reporter outside chroot. The latter doesn't have the job end time.
To fix this problem, the reporter outside chroot should reload the stats
proto to get the job end time.

Bug: 336239721
Bug: 311377497
Test: Run Pre-reboot Dexopt twice.
Change-Id: Iaf8f574087e2ebbc99aef1004d862ae2777e80c7