Revert "Update test status after D8 fix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9f3d077-d4de-4462-813c-d9de993b6191). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I94d2c31a6bd7a05ba0a3d4992618ddcc298d3361
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com