Make sure to mark classes/methods as startup.
The GetMinFirstSaveMs() which depends on the system property
[dalvik.vm.ps-min-first-save-ms](https://source.android.com/docs/core/runtime/configure#:~:text=dalvik.vm.ps%2Dmin%2Dfirst%2Dsave%2Dms) should only impact the waiting
time to dump the profile to disk, but it shouldn't impact marking
classes/methods as startup.

After this change, even if the system property `dalvik.vm.ps-min-first-save-ms` was configured with a high waiting time, it will normally mark
classes/methods as startup after either calling reportFullyDrawn or
reaching the 5 seconds limit, but it still wait for the stated waiting time before dumping the profile to disk.

Bug: 319911625
Test: Running the BaselineProfileRuleTest `adb shell am instrument -w -e androidx.benchmark.targetPackageName androidx.benchmark.integration.macrobenchmark.target -e class androidx.benchmark.integration.macrobenchmark.BaselineProfileRuleTest androidx.benchmark.integration.macrobenchmark/androidx.test.runner.AndroidJUnitRunner`
Change-Id: Ic9e727f449fe7feb858de6d2f55460804b774cff