Make sure to mark classes/methods as startup.
The GetMinFirstSaveMs() which depends on the system property
`dalvik.vm.ps-min-first-save-ms` should only impact the waiting
time to dump the profile to disk, but it shouldn't impact marking
classes/methods as startup.

With this change, even if the system property `dalvik.vm.ps-min-first-save-ms`
was configured with a high waiting time, we will still mark classes/methods
as startup after either calling reportFullyDrawn or reatching the 5 seconds
limit.

With this change if the `dalvik.vm.ps-min-first-save-ms` is less than
`GetMinSavePeriodMs` then it will wait for `GetMinSavePeriodMs`to be
dumped to `cur` profile, this is still expected since
`dalvik.vm.ps-min-first-save-ms` is the min time to dump the first
profile.

Bug: 319911625
Test: Running the BaselineProfileRuleTest `adb shell am instrument -w -e androidx.benchmark.targetPackageName androidx.benchmark.integration.macrobenchmark.target -e class androidx.benchmark.integration.macrobenchmark.BaselineProfileRuleTest androidx.benchmark.integration.macrobenchmark/androidx.test.runner.AndroidJUnitRunner`
Change-Id: Ic9e727f449fe7feb858de6d2f55460804b774cff