Revert "Only unmap snapshots if we have mapped snapshots."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/823bc622-dd5d-452a-a40e-6e26a48943f2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 342693574
Change-Id: I851284a8bd0e810a1459ecaf90433103a91c7f73
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com