Revert "Clean up descriptor and member name comparisons."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2407debc-6086-495c-a7a3-77e500dbddbc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I82b94de924ee91425dfd49356742b74e5500b2bd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com