Fix the way we detect the GC type from the command line.
The previous code was looking for either the "-Xgc:CMC" or "-Xgc:CC" strings, but this is not always the case. For example, if the command line is "-Xgc:CMC ... -Xgc:CC", the code would incorrectly detect the GC type as CMC, even though the last flag is CC and the last flag takes priority.

The new code uses the rfind() function to find the last occurrence of "-Xgc:" in the command line, and then uses substr() to get the substring that follows it. This substring will be either "CMC" or "CC", depending on the GC type.
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:bae81acd5431bc05c0ca8447f9fbde44e5734cce)
Merged-In: Id103619fdad5355a5206b95721be0013773bd4f5
Change-Id: Id103619fdad5355a5206b95721be0013773bd4f5