Revert "Avoid transaction records for all new objects."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5fe5d045-d34f-421b-8163-8bb0399314d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3507aa048aeae2c6b9631dc7ff77126975defaa5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com