Revert "Rewrite how we identify hot methods."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d729e15-5ebc-4a8d-8fed-3c620fd3f93c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id456b46a34dc6ad750441bedd3cc9e333706e78e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com