Revert "Rewrite how we identify hot methods."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/70e36835-61fa-48e3-95dd-81d09ed7300b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib6ad9dd32023d2ce5e8c6bf24cae0213acd6eb20
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com