Revert "Avoid transaction records for all new objects."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4f206867-1cc6-4931-bd66-162a66d6c401). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibace1aa9ca64e0e7b90ff442ba811408ef35f89e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com