Revert "Rewrite how we identify hot methods."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f936120-f1ad-41b6-956c-8945cea86d65). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I81a77eadda7ba92ebc154ec82a4fd34890424fff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com