Revert "Addd check for protect AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f936120-f1ad-41b6-956c-8945cea86d65). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I281cd0a1546e50d985122cfea47f53b2f55961e8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com