Revert "Rewrite how we identify hot methods."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/26a9f684-404a-4992-8c38-edb79c4ab50f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icbc2350df5680d9dcc8cd7973faf44d682b1b93f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com