Revert "Process tracing buffers in the correct order"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6f390664-ecd6-40f6-b14d-cba86ca309ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia77c47afb1fbe65b6a5c7421aa14b2a741f95270
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com