Revert "Add `pm art pr-dexopt-job` shell command."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0587a4fa-dfb8-4f2c-af36-4f018b4a6994). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I58757cad8c75c0156359d86e74593a928a574b8b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com