Revert "Interface methods can also call protected methods."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/258c9e8c-5ebb-458c-bc12-e3eeafe2bbf2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia93a7c054e16950240b9b8a90029f9a45ade0168
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com