Revert "Use process groups to manage artd's subprocesses."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/91243d2d-2db2-4041-a90c-97e123f89902). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4d44fa5c974493d8358821e494a4f1f5c7ee38f5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com