Revert "riscv: Expand BitwiseNegatedRight to riscv64, optimize"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/34a7562d-aecb-4f2b-a5d7-c810f39daf61). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic00e8b37362cc09f9adeeeecfc7701e2ce2c7096
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com