Revert "Always set page agnostic flag."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b176771d-2f67-4e8c-8669-e5ea18a1ace0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61db07d1cf6335e1c66f27a10decb74b97c0831d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com