Revert "Always set page agnostic flag."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2063cb38-5c7c-41a1-bd04-f6c54d10c491). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6c4389882ba91127d4da626d0ab5fae2a33a2095
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com