Revert "Always set page agnostic flag."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ca8b0eb-15d7-427c-ae4d-f36c488e9e54). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I07da1e167b9b374403f1735def8fdea9681958dd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com