Reduce 2029-contended-monitors iteration count
... for debuggable runs to prevent timeouts.

This required some other minor changes since we determine whether we
are debuggable only at run-time, and some of the long lock hold times
are no longer feasible with loower iteration counts. We want to keep
the output constant, so we reduced iteration consistently for all
configurations.

Unfortunately, by introducing a dependency on libarttest, this also
breaks the atest support. Disabled that for now.

Test: testrunner.py --host -b -t 2029-contended-monitors
Bug: 343887569
Change-Id: Iedd38f993524d34ff83e59e0a6543231ca554b4c