Drop native static libs from the SDK.
They don't have a stable ABI that allows them to be dropped as
prebuilts into older platform releases. In particular,
libdexfile_support depends on STL, which got updated in VIC and doesn't
have the same symbols anymore.

Platforms that get prebuilt SDKs after this change will need a patch
like https://r.android.com/3138054 to enable building the corresponding
libraries from source, even when the rest of the SDK is prebuilt.

The static libraries were added with the goal of unbundling ART sources
from the platform altogether. That effort has since been abandoned, and
it's not an issue to build these libraries from source in the ART
version that's in the platform and use them with an update prebuilt
module, because all module APIs still use stubs.

Also drop some other shared libraries that were added for the same
purpose - they're all intended to be used by code outside the module
and are hence better built from source with that code (typically
platform).

Test: packages/modules/common/build/mainline_modules_sdks.sh \
        --build-release=UpsideDownCake
  Check that the build SDK doesn't have any of the dropped libs.
Bug: 346746614
Change-Id: I682e389f508c69bce1b16fce3c65e2e8836d3cb3