Revert "Use the correct lock in TraceWriter::WriteToFile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/57ea6c0b-cf15-4309-b673-b594aa256aa8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I07ffd1caf93b3b709036f2775e56e3f996fce2ea
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com