Revert "Only increase the hotness for UI thread after startup."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c6efce4-3ca9-48cc-bf61-67dbe562c2c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8d46d53a9d6fe8ffb8851779c5f13cf2edbaa8c1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com