Revert "Only increase the hotness for UI thread after startup."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4f0ba0c-b98c-4a51-8212-00d43cb1e505). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7e49c8881bd6307e7b4662972c139de2fb5cbc49
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com