Revert "Handle null source file when providing method information"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c2d05029-18c6-409e-a434-a876563d4b63). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic480dfa621ea3fc73a8cd004360bc2e7e185fb71
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com