Revert "Only increase the hotness for UI thread after startup."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b146eeb-c341-41f1-a449-426bbf2cc803). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I96096e4e78ab5ef83befa4c4970a51d139afc633
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com