Revert "Only increase the hotness for UI thread after startup."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe1c635a-c16a-4ed3-996f-03034f92aa0d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie337ee2bd89c48f9b4eb6156001f6a47ff07e1e8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com