Remove missing nocache output of art gtests
art gtest declare an implicit output <file>-nocache which never gets
created. This is done to force the tests to run on every invocation.
This approach is not compatible with ninja's `missingoutfile=err`, which
is being enforced in https://r.android.com/3141624.

This CL drops the nocache file. The test output is pinned to an old
timestamp. This makes it perpetually stale, and therefore run in every
single invocation.

Test: m out/host/linux-x86/obj/PACKAGING/art-host-gtest_intermediates/art_libdexfile_external_tests32.xml
Change-Id: I2f648f0bcbfc7c70c6576ff575337987ebda535c