Remove missing nocache output of art gtests
art gtest declare an implicit output <file>-nocache which never gets
created. This is done to force the tests to run on every invocation.
This approach is not compatible with ninja's `missingoutfile=err`, which
is being enforced in https://r.android.com/3141624.

This CL drops the nocache file. art_gtest will continue to run on every
invocation since gtest_rule is a phony target

Test: m test-art-host-gtest-art_libdexfile_external_tests32
Test: NINJA_ARGS="-n -d explain" m
test-art-host-gtest-art_libdexfile_external_tests32 # dirty, so it will
run again

Change-Id: I2f648f0bcbfc7c70c6576ff575337987ebda535c