Revert "Only increase the hotness for UI thread after startup."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ed795c6-ab2c-4db0-bdf3-5f7d6ac5c9b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I911c3fee1b74c43ae2d91499e40c52ce558ab6f3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com