Revert "Regenerate ART test files (2024-06-20)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/746928ce-0276-437f-92d2-2494e318b48a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ica2323e238b70f69e19d1266a4f22b40924ef2ab
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com