Revert "Honor the Lint baseline file in `regen-test-files`, if p..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/746928ce-0276-437f-92d2-2494e318b48a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2a32a138e0bde382b6be346f596ebb91be4e2cda
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com