Revert "Ensure tests are both or neither in CTS and MCTS in `reg..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/746928ce-0276-437f-92d2-2494e318b48a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idce2083866f1c2a73487f004e06639fafd458bce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com