Revert "Regenerate ART test files (2024-06-20)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/746928ce-0276-437f-92d2-2494e318b48a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I20d64c54b5d814e691e44f4087f6cba47ddef60d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com