Revert "Ensure ART gtests are both or neither in CTS and MCTS."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/746928ce-0276-437f-92d2-2494e318b48a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I02d63d26b0198827c4f0dd2093c49f2f6dab7e99
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com