Revert "Add 2275-pthread-name test for Thread.setName"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8c3a77b3-4c01-450c-8225-91248ea9622a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4b549678c1c2c812c71569fa67ac0c760d6adbdd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com