Support mixed mode autovectorization.
This CL brings support for mixed mode vectorization - when a graph
could have some loops vectorized in predicated mode and some others
vectorized in traditional mode.

The vectorizer will try to vectorize each loop in traditional mode
(e.g. using arm64 NEON); if it fails to do so and if the target
supports predicated vectorization mode (e.g. arm64 target with
SVE), the vectorizer will attempt to vectorize the loop in
predicated mode. This heuristic of which mode to use, could be
extended for each target in the future if needed.

Original author: Artem Serov <Artem.Serov@linaro.org>
Co-Authored-By: Konstantin Baladurin <Konstantin.Baladurin@arm.com>

Test: ./art/test/testrunner/testrunner.py --host --optimizing --jit
Test: ./art/test/testrunner/testrunner.py --target --optimizing --jit
Test: m test-art-host-gtest
Test: run-gtests.sh

Test: target tests on arm64 with SVE (for details see
      art/test/README.arm_fvp).

Change-Id: I6dae5a4ae8fcf0d0812e64e80ace4149247b1eec