Revert "[DO NOT MERGE] Add mcts tags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a865f845-912d-4186-9d55-147c9fd4fee6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I09c6d7fcf4d567426d8b82d7ea4c4829c7a81cad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com