Revert "Don't add compile task  if we already have optimized code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7b1e4984-45c0-4846-9820-917a80e83429). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I02411bad4377b0ab5388d9d83674357d0f46cec8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com