Revert "Don't add compile task  if we already have optimized code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ac25770f-31b3-49e5-a395-12ad21898b5d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2eee048d25e5ec212340f6bf9c69b2d2f79953f0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com