Revert "Add support for multiple dirty-image-objects files"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e72162f-6614-4551-a9e4-7ebf98fcba00). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifc47a4ad85a1c712d3f4ac6ff11e4d4c05786ed3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com