Revert "Add support for multiple dirty-image-objects files"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/39fb892f-5f66-40e6-ada5-cfd52b27d713). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I286de7da7401ab2345a7f923e42a07bc377936b5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com