Revert "Offload `onStartJob` and `onStopJob` calls from the main..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/984ae244-ef15-47bf-8cbb-6ca4bbe28efe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id9d32ae43de975a37be68231c4e9b27b1e63bfca
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com