Revert "Use atomics for find_array_class_cache_"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/40c365a6-94a2-456d-84d3-a5eec096bbd1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1601ef25ca241d1084e22986977857f9d7a73c37
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com