Revert "Reland "Rewrite how we identify hot methods.""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ecc53bf-8da0-4dc6-b035-b0d115fdd3b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I31a7ea8d224cc71fb65448028b3fdbc99d84da52
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com