Revert "Update InductionVarRange::Replace to match more cases"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f7c4c9a-5419-4fd9-b6b1-3e97fb09d397). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I56a48bfb512ea18c4b4d91423a777611dd58c71a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com