Revert "hprof: output header size of arrays"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2cbff312-91c2-476a-8667-a07a83489c0d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9f0105f0e4cd118bbb9b4a74d0d5c3b103c1cea1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com