Revert "hprof: output header size of arrays"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd16140a-0469-4594-9ddf-fc20cc47d0eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I31468ef1740340f766c8182dbec0421d5835f9e6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com