Revert "Remove threaded-mode related code from CMC GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/29bb93e7-f87a-4e58-94b1-108f4f3e77f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2bfba9d93ba9495fdf88005ea93985d61d4f7861
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com